pyproject.toml
[tool.poetry]
name = "project-01"
version = "0.1.0"
description = ""
authors = ["Park Jihong <elan5027@naver.com>"]
readme = "README.md"
packages = [{include = "project_01"}]

[tool.poetry.dependencies]
python = "^3.8"
django = "^4.2"
pillow = "^9.5.0"
djangorestframework = "^3.14.0"
djangorestframework-simplejwt = "^5.2.2"
django-cors-headers = "^3.14.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
